pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]
classifiers = [
    "Programming Language :: Python",
]

[tool.poetry.dependencies]
python = "^3.11"
pytest-cov = "^4.1.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
